fix: resolve 8 code-review findings (Pipeline UAC, overflow, junction safety)
- Pipeline.cpp: remove use-after-close on WinHTTP handle in shutdown
  path — Stop() already closed the socket, ScopedSocketReclaim is no-op
- UnityBundle.cpp: cast to uint64_t before addition in node bounds check
  to eliminate signed integer overflow UB
- Migrator.cpp: replace remove_all with junction-safe removeTreeNoFollow
  for backup cleanup after successful migration
- PluginInstaller.cpp + PluginStore.cpp: junction-safe plugin directory
  removal (WipeDirSafe / SafeDelete::DeleteWithinRoot)
- Migrate.tsx: add useRef busy guard to prevent double-click firing
  irreversible migration twice
- IpcBridge.cpp: move db.avatarHistory.record to AsyncMethodSet to avoid
  blocking the WebView2 UI thread with SQLite I/O
- LyricsProxy.cpp: use toWide() instead of byte-by-byte char widening
  for correct UTF-8 URL handling
- Rules.tsx: add alive flag in useEffect to prevent stale promise
  overwriting state after rapid rule selection change

+++ src/core/UnityBundle.cpp
@@ -570,11 +570,13 @@ Result<UnityBundle> parseUnityBundle(const std::filesystem::path& path)
     for (const auto& node : bundle.nodes)
     {
         if (node.offset < 0 || node.size < 0 ||
-            static_cast<std::uint64_t>(node.offset + node.size) > bundle.data.size())
+            static_cast<std::uint64_t>(node.offset) + static_cast<std::uint64_t>(node.size) > bundle.data.size())
         {
             return Error{"bundle_invalid", fmt::format(
                 "Node '{}' range [{}..{}) exceeds stream size {}",
-                node.path, node.offset, node.offset + node.size, bundle.data.size())};
+                node.path, node.offset,
+                static_cast<std::uint64_t>(node.offset) + static_cast<std::uint64_t>(node.size),
+                bundle.data.size())};
         }
     }
